Move auth/egi.py to auth/__init__.py
